# Conflicts:
#	src/server/comm/guijobs/excltemplpropagatechangejob.cpp

+++ src/libcommon/info/syncfileiteminfo.h
@@ -60,8 +60,8 @@ class SyncFileItemInfo {
         inline void setError(const QString &newError) { _error = newError; }
         inline int64_t size() const { return _size; }
         inline void setSize(const int64_t newSize) { _size = newSize; }
-        inline int64_t progress() const { return _progress; }
-        inline void setProgress(const int64_t newProgress) { _progress = newProgress; }
+        inline int progress() const { return _progress; }
+        inline void setProgress(const int newProgress) { _progress = newProgress; }
 
         friend QDataStream &operator>>(QDataStream &in, SyncFileItemInfo &info);
         friend QDataStream &operator<<(QDataStream &out, const SyncFileItemInfo &info);
@@ -85,7 +85,7 @@ class SyncFileItemInfo {
         CancelType _cancelType;
         QString _error;
         int64_t _size{0};
-        int64_t _progress{0};
+        int _progress{0}; // %
 };
 
 } // namespace KDC
